apps/api/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "scaffold-api"
version = "0.1.0"
description = "Tornado + GraphQL API scaffold"
requires-python = ">=3.13,<3.14"
dependencies = [
  "alembic>=1.14.1,<2.0.0",
  "aiosqlite>=0.21.0,<1.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "cloud-sql-python-connector[asyncpg]>=1.18.5,<2.0.0",
  "cryptography>=46.0.5,<47.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "httpx>=0.28.1,<1.0.0",
  "pydantic>=2.11.0,<3.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "sqlalchemy[asyncio]>=2.0.38,<3.0.0",
  "strawberry-graphql>=0.261.3,<1.0.0",
  "structlog>=25.1.0,<26.0.0",
  "tornado>=6.4.2,<7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[dependency-groups]
dev = [
  "mypy>=1.15.0,<2.0.0",
  "pip-audit>=2.9.0,<3.0.0",
  "pytest>=8.3.5,<9.0.0",
  "pytest-asyncio>=0.25.3,<1.0.0",
  "ruff>=0.11.0,<1.0.0",
  "types-PyYAML>=6.0.12.20241230,<7.0.0",
]
llm_local = [
  "accelerate>=1.10.0,<2.0.0",
  "huggingface-hub>=1.3.0,<2.0.0",
  "torch>=2.7.0,<3.0.0",
  "transformers>=5.3.0,<6.0.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["huggingface_hub", "huggingface_hub.errors", "torch", "transformers"]
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
